Gemini/requirements.txt
python-dotenv
streamlit
google.generativeai
Pillow